Move OfficeDashboard component
